Public POI upload in slow network conditions #1407 - Added UI indication and offline monitoring

+++ IsraelHiking.Web/src/application/services/poi.service.ts
@@ -18,6 +18,7 @@ import { GeoJsonParser } from "./geojson.parser";
 import { MapService } from "./map.service";
 import { ToastService } from "./toast.service";
 import { FileService } from "./file.service";
+import { ConnectionService } from "./connection.service";
 import { NgRedux, select } from "../reducers/infra/ng-redux.module";
 import { AddToPoiQueueAction, RemoveFromPoiQueueAction, SetOfflinePoisLastModifiedDateAction } from "../reducers/offline.reducer";
 import { SetCategoriesGroupVisibilityAction, AddCategoryAction } from "../reducers/layers.reducer";
@@ -67,6 +68,7 @@ export class PoiService {
     private poisCache: GeoJSON.Feature[];
     private poisGeojson: GeoJSON.FeatureCollection<GeoJSON.Point>;
     private miniSearch: MiniSearch;
+    private queueIsProcessing: boolean;
 
     public poiGeojsonFiltered: GeoJSON.FeatureCollection<GeoJSON.Point>;
     public poisChanged: EventEmitter<void>;
@@ -92,10 +94,12 @@ export class PoiService {
                 private readonly toastService: ToastService,
                 private readonly mapService: MapService,
                 private readonly fileService: FileService,
+                private readonly connectionService: ConnectionService,
                 private readonly ngRedux: NgRedux<ApplicationState>
     ) {
         this.poisCache = [];
         this.poisChanged = new EventEmitter();
+        this.queueIsProcessing = false;
 
         this.poiGeojsonFiltered = {
             type: "FeatureCollection",
@@ -144,32 +148,49 @@ export class PoiService {
                     });
                 });
         }
-        this.uploadPoiQueue$.subscribe(async (items: string[]) => {
-            if (items.length === 0) {
-                this.loggingService.info(`[POIs] Upload queue changed and now it is empty`);
-                return;
-            }
-            let firstItemId = items[0];
-            this.loggingService.info(`[POIs] Upload queue changed, items in queue: ${items.length}, first item id: ${firstItemId}`);
-
-            let feature = await this.databaseService.getPoiFromUploadQueue(firstItemId);
-            try {
-                let postAddress = Urls.poi + "?language=" + this.resources.getCurrentLanguageCodeSimplified();
-                let putAddress = Urls.poi + feature.properties.poiId + "?language=" + this.resources.getCurrentLanguageCodeSimplified();
-                let poi = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(feature.properties.poiId)
-                    ? await this.httpClient.post(postAddress, feature).pipe(timeout(60000)).toPromise() as GeoJSON.Feature
-                    : await this.httpClient.put(putAddress, feature).pipe(timeout(60000)).toPromise() as GeoJSON.Feature;
-    
-                this.loggingService.info(`[POIs] Uploaded feature with id: ${firstItemId}, removing from upload queue`);
-                if (this.runningContextService.isCordova) {
-                    this.databaseService.storePois([poi]);
-                }
-                this.databaseService.removePoiFromUploadQueue(firstItemId);
-                this.ngRedux.dispatch(new RemoveFromPoiQueueAction({featureId: firstItemId}));
-            } catch (ex) {
-                this.loggingService.error(`[POI] Failed to upload feature with id: ${firstItemId}, ${ex.message}`);
+        this.uploadPoiQueue$.subscribe((items: string[]) => this.handleUploadQueueChanges(items));
+        this.connectionService.monitor(false).subscribe(state => {
+            this.loggingService.info(`[POIs] Connection status changed to: ${state.hasInternetAccess}`);
+            if (state.hasInternetAccess && this.ngRedux.getState().offlineState.uploadPoiQueue.length > 0) {
+                this.handleUploadQueueChanges(this.ngRedux.getState().offlineState.uploadPoiQueue);
             }
         });
+
+    }
+
+    private async handleUploadQueueChanges(items: string[]) {
+        if (items.length === 0) {
+            this.loggingService.info(`[POIs] Upload queue changed and now it is empty`);
+            return;
+        }
+        if (this.queueIsProcessing) {
+            this.loggingService.info(`[POIs] Upload queue is currently processing, ignoring changes`);
+            return;
+        }
+        this.queueIsProcessing = true;
+        let firstItemId = items[0];
+        this.loggingService.info(`[POIs] Upload queue changed, items in queue: ${items.length}, first item id: ${firstItemId}`);
+
+        let feature = await this.databaseService.getPoiFromUploadQueue(firstItemId);
+        try {
+            let postAddress = Urls.poi + "?language=" + this.resources.getCurrentLanguageCodeSimplified();
+            let putAddress = Urls.poi + feature.properties.poiId + "?language=" + this.resources.getCurrentLanguageCodeSimplified();
+            let poi = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(feature.properties.poiId)
+                ? await this.httpClient.post(postAddress, feature).pipe(timeout(60000)).toPromise() as GeoJSON.Feature
+                : await this.httpClient.put(putAddress, feature).pipe(timeout(60000)).toPromise() as GeoJSON.Feature;
+
+            this.loggingService.info(`[POIs] Uploaded feature with id: ${firstItemId}, removing from upload queue`);
+            if (this.runningContextService.isCordova) {
+                this.databaseService.storePois([poi]);
+            }
+            this.databaseService.removePoiFromUploadQueue(firstItemId);
+            this.queueIsProcessing = false;
+            this.ngRedux.dispatch(new RemoveFromPoiQueueAction({featureId: firstItemId}));
+        } catch (ex) {
+            this.loggingService.error(`[POI] Failed to upload feature with id: ${firstItemId}, ${ex.message}`);
+        } finally {
+            this.queueIsProcessing = false;
+        }
     }
 
     private async getPoisFromServer(): Promise<GeoJSON.Feature<GeoJSON.Point>[]> {
